Ecommerce_catalog_service/requirements.txt
fastapi
uvicorn[standard]
mysql-connector-python
pydantic
pydantic-settings
python-dotenv
mysqlclient
